requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0 